Refactor and consolidate authentication tests
- Removed outdated test files: auth.routes.final.test.js, auth.routes.test.js.backup, and auth.routes.test.js.disabled.
- Introduced a new test suite in auth.test.js with streamlined tests for signup, signin, logout, password reset, and email verification.
- Simplified mock implementations for authentication service and middleware.
- Enhanced error handling and response validation in tests.
- Deleted setup.js and testCleanup.js as they were no longer needed.
