+++ package.json
@@ -5,14 +5,8 @@
 	"scripts": {
 		"start": "nodemon ./bin/www",
 		"test": "jest",
-		"test:stable": "jest test/auth.routes.final.test.js test/auth.controller.test.js test/config.test.js --runInBand --forceExit",
 		"test:watch": "jest --watch",
 		"test:coverage": "jest --coverage",
-		"test:auth": "jest test/auth --runInBand --forceExit",
-		"test:auth:watch": "jest test/auth --watch",
-		"test:auth:coverage": "jest test/auth --coverage --runInBand --forceExit",
-		"test:auth:final": "jest test/auth.routes.final.test.js --runInBand --forceExit",
-		"test:auth:simple": "jest test/auth.simple.test.js --runInBand --forceExit",
 		"lint": "eslint . --ext .js,.jsx,.ts,.tsx",
 		"lint:fix": "eslint . --ext .js,.jsx,.ts,.tsx --fix"
 	},
